Require a listener registration before sending
If sendToLocalListeners() is invoked before a listener has registered, the function will emit a packet to 0.0.0.0 port 0, since the variables remote and localPort have not yet been initialized.

In some reported circumstances this causes problems with the DHCP server of certain cable-internet routers which sees the emitted packet having the device MAC address and the default address 192.168.0.51 (at https://github.com/telldus/tellstick-net/blob/master/firmware/TCPIPConfig.h#L28 ). The DHCP server then assumes that the device is already assigned an adress and does not hand it a DHCP address.

The proposed patch is to add a flag, "listenerActive", which is set to true once at least one listener has registered. If no listener has registered, no UDP packet will be put out, and thus the device will receive the DHCP lease as usual.